Fix removing a file
When a file is removed, the directory in which it is gets an on_modified
signal but there is no file modified per say since it was removed.
So we no just check if we can find the modified file and if not (ie it
was deleted) we keep on going.