Rework the logic to update the repo
- First of all with this commit we use the same code to update the repo
when running in daemon mode or in single-run.
With this change, we no longer end up with the risk of adding to the repo
file that should be excluded via the .gitignore
- We used to do: git stash, git pull --rebase, git stash pop before doing
the commit and the git push. But that meant that we were changing the
file on disk quite frequently and frankly for no good reason when (most
cases) there was no changes to pull. We now are doing git fetch, checking
the latest commit locally and in the remote and only stash to pull --rebase
if the commits are different.